+++ genie-client/src/main/java/com/netflix/genie/client/apis/TokenService.java
@@ -17,7 +17,7 @@
  */
 package com.netflix.genie.client.apis;
 
-import com.netflix.genie.client.security.oauth.AccessToken;
+import com.netflix.genie.client.security.oauth2.AccessToken;
 import retrofit2.Call;
 import retrofit2.http.FieldMap;
 import retrofit2.http.FormUrlEncoded;
