+++ genie-web/build.gradle
@@ -40,6 +40,7 @@ dependencies {
     // Swagger libs
     compile("io.springfox:springfox-swagger2:${springfox_version}")
     compile("io.springfox:springfox-swagger-ui:${springfox_version}")
+    compile("io.springfox:springfox-bean-validators:${springfox_version}")
 
     /*******************************
      * Provided Dependencies
