+++ gradle.properties
@@ -38,22 +38,10 @@ spectator_version=0.30.0
 # Swagger Libraries
 springfox_version=2.4.0
 
-# WebJar Versions
-bootstrap_version=2.3.2
-datatables_version=1.10.7
-datatables_bootstrap_version=2-20120202-2
-jquery_version=2.0.2
-knockout_version=2.2.1
-lodash_version=1.3.1
-momentjs_version=2.10.3
-requirejs_version=2.1.20
-requirejs_text_version=2.0.14
-select2_version=3.4.5
-
 # Test Libraries
-dbunit_version=2.5.1
+dbunit_version=2.5.2
 jtidy_version=r938
-spring_test_dbunit_version=1.2.1
+spring_test_dbunit_version=1.3.0
 
 # Findbugs
 findbugs_annotations_version=3.0.1
